Create the Jumpup::TestCase
This makes Jumpup works with Rails 3.x and 4.x.
In the rails 3.x the minitest version is 4.x in the
rails 4.x the minitest version is 5.x. This way we can't simply define
the minitest 5 on the gemspec since this will raise a warning on rails
3.x. Because from minitest 4 to 5 the test class change from
MiniTest::Unit::TestCase to Minitest::Test.

To fix this we create Jumpup::TestCase that uses the correct test case
class.